requirements.txt
aliyun-python-sdk-core==2.13.9
Django==1.11.7
django-cors-headers==3.1.0
django-redis==4.10.0
djangorestframework==3.9.4
djangorestframework-stubs==1.0.0
jmespath==0.9.4
PyMySQL==0.9.3
pytz==2019.1
redis==3.2.1
sqlparse==0.3.0
uWSGI==2.0.18
